Add benchmark for search()
